break out the sample data functions
