+++ src/Celeste.jl
@@ -7,6 +7,7 @@ require(joinpath(Pkg.dir("Celeste"), "src", "SDSS.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "ElboDeriv.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "OptimizeElbo.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "Synthetic.jl"))
+require(joinpath(Pkg.dir("Celeste"), "src", "SampleData.jl"))
 
 import SDSS
 import ElboDeriv
